# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "REST API E-Learning Platform",
        "version": "1.0"
    },
    "paths": {
        "127.0.0.1:8000/api/v1/categories": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get all categories",
                "operationId": "e6ed4939b3524a2011f17073559f3a9c",
                "responses": {
                    "200": {
                        "description": "A Category endpoint"
                    }
                }
            }
        },
        "/api/courses": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get all courses",
                "operationId": "ddca64e00ac1eac83c478a88f7f78de2",
                "responses": {
                    "200": {
                        "description": "A list of courses"
                    }
                }
            },
            "post": {
                "tags": [
                    "courses"
                ],
                "summary": "Create a new course",
                "operationId": "ab90953dc6f70ea571640a50118f7279",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "category_id",
                                    "mentor_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Mathematics 101"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "A basic introduction to mathematics."
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "mentor_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Course created successfully"
                    }
                }
            }
        },
        "/api/v1/categories": {
            "post": {
                "tags": [
                    "categories"
                ],
                "summary": "Create a new category",
                "operationId": "f9171669d626697cd543f9f421cff718",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Electronics"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1,
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Category created successfully"
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "categories"
                ],
                "summary": "Get category by ID",
                "operationId": "11d2d45398d538612b1a4156d210652e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        },
                                        "category_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "categories"
                ],
                "summary": "Updates a category",
                "operationId": "8edd529bf8c5d2c024f479805defe8b6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Advanced Mathematics"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 9
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "category  updated successfully"
                    },
                    "404": {
                        "description": "Category not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "categories"
                ],
                "summary": "Delete a category",
                "operationId": "91c938105382f9b2d833cb18706e6694",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the category to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Category deleted successfully"
                    }
                }
            }
        },
        "/api/courses/{id}": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get a course by ID",
                "operationId": "03ec4e1d2fffa35c49c36c6f769b3011",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "description": {
                                            "type": "string"
                                        },
                                        "category_id": {
                                            "type": "integer"
                                        },
                                        "mentor_id": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "courses"
                ],
                "summary": "Update a course",
                "operationId": "b5d3fffeae42e27e112f51efe9dc2b63",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "category_id",
                                    "mentor_id"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Advanced Mathematics"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "An in-depth study of advanced mathematics."
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 9
                                    },
                                    "mentor_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Course updated successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "courses"
                ],
                "summary": "Delete a course",
                "operationId": "3b29a86a2bc0cefa561d362578728ad3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the course to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Course deleted successfully"
                    },
                    "404": {
                        "description": "Course not found"
                    }
                }
            }
        },
        "127.0.0.1:8000/api/v1/tags": {
            "get": {
                "tags": [
                    "tags"
                ],
                "summary": "Get all tags",
                "operationId": "e25640035a20b34adf2df431b0e29055",
                "responses": {
                    "200": {
                        "description": "A Tag endpoint"
                    }
                }
            }
        },
        "/api/v1/tags": {
            "post": {
                "tags": [
                    "tags"
                ],
                "summary": "Create a new tag",
                "operationId": "894e18117a0bd232799659faff527935",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Electronics"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tag created successfully"
                    }
                }
            }
        },
        "/api/v1/tags/{id}": {
            "get": {
                "tags": [
                    "tags"
                ],
                "summary": "Get tag by ID",
                "operationId": "6d53cd01174aa367f86c6569c3bfb00b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tag",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "name": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tag not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "tags"
                ],
                "summary": "Updates a tag",
                "operationId": "b506bdc5e3250ee975ea21e92428d15d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tag to update",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Advanced Mathematics"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tag  updated successfully"
                    },
                    "404": {
                        "description": "Tag not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "tags"
                ],
                "summary": "Delete a tag",
                "operationId": "fdf9617754d2edb3b4d885a6a5f814a8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the tag to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tag deleted successfully"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "categories",
            "description": "categories"
        },
        {
            "name": "courses",
            "description": "courses"
        },
        {
            "name": "tags",
            "description": "tags"
        }
    ]
}